requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
docutils==0.16
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Kivy==2.0.0
Kivy-Garden==0.1.4
lxml==4.6.2
MarkupSafe==1.1.1
Pygments==2.8.0
pymongo==3.11.3
python-dotenv==0.15.0
requests==2.25.1
selenium==3.141.0
soupsieve==2.1
stripe==2.56.0
urllib3==1.24.3
Werkzeug==1.0.1
